% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_sequences.R
\name{best_placement}
\alias{best_placement}
\title{Best supported placement.}
\usage{
best_placement(place)
}
\arguments{
\item{place}{A data frame, the output of \code{classify_sequences}.}
}
\value{
A reduced data frame with best supported classification for each query sequence.
(ML classification for EPA, MAP for pplacer.)
}
\description{
Finds the maximum likelihood or maximum aposteriori placement of each sequence.
}
