% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_jplace.R
\name{classify_jplace}
\alias{classify_jplace}
\title{Classification of \code{jplace} object.}
\usage{
classify_jplace(jp, clade, ancestral = TRUE)
}
\arguments{
\item{jp}{An object of class \code{jplace}.}

\item{clade}{A data frame defining classification of tips (1st column) to clades / bipartitions
(2nd column). The clades can be mutually nested.}

\item{ancestral}{Logical, whether to classify ancestral branches; The default option is \code{TRUE},
but it is relevant only if the tree is rooted and defined clades are comprehensive
and non-overlapping (otherwise switched to \code{FALSE} and issues a warning).}
}
\value{
An amended \code{jplace} object with additional components \code{clade} and \code{position}
in the \code{placements} data frame, which indicate classification of the branches.
}
\description{
Classifies tree edges with any phylogenetic placements to clades (or bipartitions).
}
