% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_sequences.R
\name{classify_sequences}
\alias{classify_sequences}
\title{Classification of sequences.}
\usage{
classify_sequences(jp, keep = NULL, probthr = 0, cumprobthr = 1)
}
\arguments{
\item{jp}{An object of class \code{jplace} pre-treated by \code{classify_jplace},
i.e., containing classification to clades / bipartitions.}

\item{keep}{Numeric, no. of placements per clade to keep, corresponds to 'epa-­keep-­placements' in RAxML.
\code{keep=1} means the maximum likelihood placement, while \code{keep=NULL} (default) means all placements.}

\item{probthr}{Numeric, likelihood weight threshold, corresponds to 'epa­-prob­-threshold' in RAxML.}

\item{cumprobthr}{Numeric, cummulative likelihood weight threshold, corresponds to
'epa­-accumulated­-threshold'in RAxML.}
}
\value{
A data frame with components:
\code{id} with labels of query sequences;
\code{clade} with labels of clade / bipartition;
\code{position} which is either \code{"crown"} or \code{"stem"};
\code{probability} with probability of classification, i.e., sum of probabilities over the branches
belonging to the clade / bipartition. In this context, 'probability' means either posterior probability
or relative likelihood weight, i.e., a quantity which sums up to unity over all branches of the tree.
\code{n_branch} which is the number of branches which contributed to the sum.
}
\description{
Classifies sequences to clades (or bipartitions).
}
