% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jplace.R
\name{make_plotlabels}
\alias{make_plotlabels}
\title{Branch labels.}
\usage{
make_plotlabels(jp, classification, sep = "\\n")
}
\arguments{
\item{jp}{An object of class \code{jplace}.}

\item{classification}{A data frame, output of \code{classify_sequences}.}

\item{sep}{A character string separating labels of query sequences placed on the same branch.}
}
\value{
A character vector with concatenated id's of query sequences most probably placed on given branch.
The vector elements are named by branch labels from \code{x$tree$edge.label}.
}
\description{
Makes labels indicating which query sequences are most probably placed on given branches.
}
