% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jplace.R
\name{plot.jplace}
\alias{plot.jplace}
\title{Plotting phylogenetic placements.}
\usage{
\method{plot}{jplace}(
  x,
  subset = NULL,
  keep,
  prob = 1,
  type = "radial",
  col = 3,
  bg = 1,
  file,
  ...
)
}
\arguments{
\item{x}{An object of class \code{jplace}.}

\item{subset}{Character vector with sequence names.}

\item{keep}{Numeric, how many best placements to keep for display.}

\item{prob}{Numeric, ummulative probability threshold up to which the placements are displayed.}

\item{type}{Character string indicating the type of tree, reasonable is \code{"radial"} (default) and \code{"unrooted"}.}

\item{col}{Color to highlight branches with placements.}

\item{bg}{Color of background branches.}

\item{file}{Optional character string, the name of .pdf file into which the plot is saved.}

\item{...}{Other parameters of \link[ape:plot.phylo]{ape::plot.phylo}.}
}
\description{
A plot method for class \code{jplace}, it highlights branches with placements of query sequences.
}
