% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jplace.R
\name{read_jplace}
\alias{read_jplace}
\title{Reading .jplace file.}
\usage{
read_jplace(jp)
}
\arguments{
\item{jp}{Character, the name of '.jplace' output of EPA or pplacer.}
}
\value{
An object of class \code{jplace}, which is a list with components:
\code{tree}, a \code{phylo} object with the unrooted reference tree used by EPA;
\code{placements}, a data.frame with EPA output, i.e., list of possible placements
of query sequences into the tree and their supports (likelihoods or posterior probabilities);
and \code{software}, a character string indicating the software used (either \code{"pplacer"} or \code{"raxml"}).
}
\description{
Reads an output of RAxML's EPA or pplacer in '.jplace' format.
}
