% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_jplace.R
\name{root_jplace}
\alias{root_jplace}
\title{Rooting \code{jplace} object.}
\usage{
root_jplace(jp, outgroup)
}
\arguments{
\item{jp}{An object of class \code{jplace}.}

\item{outgroup}{A character vector listing outgroup sequences or their unique identifier(s),
e.g., a character string "outgroup".}
}
\value{
A modified \code{jplace} object, i.e., its \code{tree} is rooted and \code{placements$edge_num} modified.
}
\description{
Roots the tree & adjusts edge labelling accordingly.
}
