% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.delim.R
\name{write.delim}
\alias{write.delim}
\title{Writing tab-delimited file.}
\usage{
write.delim(x, file, row.names = FALSE, col.names = TRUE, quote = FALSE)
}
\arguments{
\item{x}{A matrix or data frame to be written.}

\item{file}{Name of the output file.}

\item{row.names}{Logical, whether to use row names (default is FALSE).}

\item{col.names}{Logical, whether to use column names (default is TRUE).}

\item{quote}{Logical or numeric, whether to enclose field contents into quotes (or in which columns).}

\item{sep}{the field separator string.}
}
\description{
A wrapper for \link[utils:write.table]{utils::write.table} writing a tab-delimited text file. The counterpart of \link[utils:read.table]{utils::read.delim}.
}
